% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{abind_time}
\alias{abind_time}
\title{Concatenate two 3-D arrays along the third dimension (time)}
\usage{
abind_time(a, b)
}
\description{
Concatenate two 3-D arrays along the third dimension (time)
}
\keyword{internal}
