% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind.R
\name{calculate_period_wind_exceedance_frequency}
\alias{calculate_period_wind_exceedance_frequency}
\title{Calculate monthly wind exceedance frequency}
\usage{
calculate_period_wind_exceedance_frequency(
  u10_dataset,
  v10_dataset,
  reference_period
)
}
\arguments{
\item{u10_dataset}{List for \code{u10}.}

\item{v10_dataset}{List for \code{v10}.}

\item{reference_period}{Character vector \code{c("start", "end")}.}
}
\value{
A list with \code{data} [lon x lat x periods] and \code{time}.
}
\description{
Calculate monthly wind exceedance frequency
}
