% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind.R
\name{days_above_wind_thresholds}
\alias{days_above_wind_thresholds}
\title{Calculate days with wind power above the 90th-percentile threshold}
\usage{
days_above_wind_thresholds(u10_dataset, v10_dataset, reference_period)
}
\arguments{
\item{u10_dataset}{List for \code{u10}.}

\item{v10_dataset}{List for \code{v10}.}

\item{reference_period}{Character vector \code{c("start", "end")}.}
}
\value{
A list with binary \code{data} [lon x lat x days] and \code{time}.
}
\description{
Calculate days with wind power above the 90th-percentile threshold
}
