% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{.aggregate_granularity_array}
\alias{.aggregate_granularity_array}
\title{Aggregate a 3-D climate array along the time dimension}
\usage{
.aggregate_granularity_array(data, time, granularity = "month", FUN = mean)
}
\arguments{
\item{data}{Numeric array \code{[nl x nw x nt]}.}

\item{time}{POSIXct vector of length \code{nt}.}

\item{granularity}{One of \code{"month"}, \code{"season"},
\code{"semester"}, \code{"year"}.}

\item{FUN}{Aggregation function. Default \code{mean}.}
}
\value{
A list with \code{data} (array \code{[nl x nw x nt_agg]}),
  and \code{time} (character vector of period labels, same convention
  as \code{aggregate_granularity()}).
}
\description{
Aggregate a 3-D climate array along the time dimension
}
\keyword{internal}
