% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{.merge_netcdf_files}
\alias{.merge_netcdf_files}
\title{Merge multiple yearly NetCDF files along the time dimension}
\usage{
.merge_netcdf_files(files, output, variable)
}
\arguments{
\item{files}{Character vector of NetCDF paths.}

\item{output}{Output NetCDF file.}

\item{variable}{Variable name.}
}
\description{
Merge multiple yearly NetCDF files along the time dimension
}
\keyword{internal}
