% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{.rename_mask_variable}
\alias{.rename_mask_variable}
\title{Rename the ERA5 land-sea mask variable to 'country' and drop the time dim}
\usage{
.rename_mask_variable(raw_file, mask_file)
}
\description{
Rename the ERA5 land-sea mask variable to 'country' and drop the time dim
}
\keyword{internal}
