% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_netcdf}
\alias{load_netcdf}
\title{Load a NetCDF variable as a 3D array [lon x lat x time]}
\usage{
load_netcdf(path, var_name)
}
\arguments{
\item{path}{Path to the NetCDF file.}

\item{var_name}{Name of the variable to extract.}
}
\value{
A list with fields: \code{data} (3D array), \code{lon}, \code{lat}, \code{time}.
}
\description{
Load a NetCDF variable as a 3D array [lon x lat x time]
}
