% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_psmsl_data}
\alias{load_psmsl_data}
\title{Load the bundled PSMSL station metadata}
\usage{
load_psmsl_data()
}
\value{
A \code{data.frame} of PSMSL tide-gauge station information.
}
\description{
Load the bundled PSMSL station metadata
}
