% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_dataframes}
\alias{merge_dataframes}
\title{Merge a list of data frames on their row-name index}
\usage{
merge_dataframes(dataframes)
}
\arguments{
\item{dataframes}{A list of \code{data.frame} objects sharing a common row-
name index.}
}
\value{
A single merged \code{data.frame}.
}
\description{
Merge a list of data frames on their row-name index
}
\keyword{internal}
