% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{resample_daily}
\alias{resample_daily}
\title{Resample a 3-D climate array to daily resolution}
\usage{
resample_daily(dataset, FUN = sum)
}
\arguments{
\item{dataset}{List returned by \code{load_component()} (may have
sub-daily time steps).}

\item{FUN}{Aggregation function: \code{sum} or \code{mean}.}
}
\value{
A new list with daily \code{time} and aggregated \code{data}.
}
\description{
Aggregates sub-daily data to daily values using the supplied function
(\code{sum} or \code{mean}).
}
\keyword{internal}
