% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\name{resample_monthly}
\alias{resample_monthly}
\title{Resample a 3-D or 1-D climate series to monthly resolution}
\usage{
resample_monthly(data, time, FUN = mean)
}
\arguments{
\item{data}{Either a 3-D array \code{[lon × lat × time]} or a numeric
vector of length \emph{T}.}

\item{time}{POSIXct vector of length \emph{T}.}

\item{FUN}{Aggregation function. Default \code{mean}.}
}
\value{
A list with \code{data} (monthly aggregates) and \code{time}
  (first day of each month as POSIXct).
}
\description{
Applies \code{FUN} within each calendar month.
}
