% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sealevel.R
\name{sealevel_clean_data}
\alias{sealevel_clean_data}
\title{Replace PSMSL sentinel values (-99999) with NA}
\usage{
sealevel_clean_data(df)
}
\arguments{
\item{df}{\code{data.frame} of sea-level measurements.}
}
\value{
Cleaned \code{data.frame}.
}
\description{
Replace PSMSL sentinel values (-99999) with NA
}
