% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind.R
\name{wind_thresholds}
\alias{wind_thresholds}
\title{Calculate wind power thresholds (90th percentile per day-of-year)}
\usage{
wind_thresholds(u10_dataset, v10_dataset, reference_period)
}
\arguments{
\item{u10_dataset}{List for \code{u10} variable.}

\item{v10_dataset}{List for \code{v10} variable.}

\item{reference_period}{Character vector \code{c("start", "end")}.}
}
\value{
A list with \code{data} [lon x lat x all_days] giving the threshold
  for each day of the full series, and \code{time}.
}
\description{
Calculate wind power thresholds (90th percentile per day-of-year)
}
