% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sealevel.R
\name{sealevel_process}
\alias{sealevel_process}
\title{Full sea-level processing pipeline}
\usage{
sealevel_process(directory, study_period, reference_period)
}
\arguments{
\item{directory}{Path to the directory with PSMSL \code{.txt} files.}

\item{study_period}{Character vector \code{c("start", "end")}.}

\item{reference_period}{Character vector \code{c("start", "end")}.}
}
\value{
A standardised \code{data.frame} of sea-level anomalies.
}
\description{
Full sea-level processing pipeline
}
