% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaci-package.R
\docType{package}
\name{xaci-package}
\alias{xaci}
\alias{xaci-package}
\title{xaci: Package xaci for Actuarial Climate Index}
\description{
Computes the Actuarial Climate Index (ACI) and its components (temperature, precipitation, drought, wind, sea level) from gridded climate data (NetCDF) and tide gauge records. Implements the methodology described in Garrido, Milhaud & Olympio (2026) for a French/European actuarial climate index, building on the American Academy of Actuaries framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/XavierMilhaud/ACI-Rpackage}
  \item Report bugs at \url{https://github.com/XavierMilhaud/ACI-Rpackage/issues}
}

}
\author{
\strong{Maintainer}: Xavier Milhaud \email{xavier.milhaud.research@gmail.com} [contributor]

Authors:
\itemize{
  \item Dabakh Kane \email{kanedabakh55@gmail.com} [contributor]
}

Other contributors:
\itemize{
  \item Esteban Mauboussin \email{esteban.mauboussin@cnp.fr} [contributor]
}

}
\keyword{internal}
